+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Optimize `u8aToHex` with direct (non-hex) conversion
+- Change all occurences of `.substr(...)` to `.substring(...)`
 
 
 ## 9.1.1 Apr 30, 2022